--- .eslintrc.json
@@ -1,24 +1,0 @@
-{
-    "env": {
-        "browser": true,
-        "es2020": true
-    },
-    "extends": [
-        "plugin:react/recommended",
-        "airbnb",
-        "airbnb/hooks",
-        "prettier",
-        "prettier/react"
-    ],
-    "parserOptions": {
-        "ecmaFeatures": {
-            "jsx": true
-        },
-        "ecmaVersion": 11,
-        "sourceType": "module"
-    },
-    "plugins": ["react"],
-    "rules": {
-        "no-underscore-dangle": "off"
-    }
-}
