--- .prettierrc.json
@@ -1,6 +1,0 @@
-{
-    "tabWidth": 4,
-    "semi": true,
-    "singleQuote": false,
-    "trailingComma": "all"
-}
