+++ README.md
@@ -4,12 +4,12 @@ This is the frontend for RiceDiscuss. It interfaces with the [backend](https://g
 
 ## Tech to be familiar with
 
--   [Apollo GraphQL](https://www.apollographql.com/docs/react/)
--   [React](https://reactjs.org/)
-    -   In particular, be familiar with the particuliarities of [Create React App](https://create-react-app.dev/)
-    -   Also, learn a little bit about [React Hooks](https://reactjs.org/docs/hooks-intro.html), especially if you're more used to traditional React classes.
--   [Styled Components](https://styled-components.com/)
--   [Material UI](https://material-ui.com/)
+- [Apollo GraphQL](https://www.apollographql.com/docs/react/)
+- [React](https://reactjs.org/)
+  - In particular, be familiar with the particuliarities of [Create React App](https://create-react-app.dev/)
+  - Also, learn a little bit about [React Hooks](https://reactjs.org/docs/hooks-intro.html), especially if you're more used to traditional React classes.
+- [Styled Components](https://styled-components.com/)
+- [Material UI](https://material-ui.com/)
 
 Furthermore, the RiceApps leadership team has a list of tutorials you can ask them for.
