+++ package.json
@@ -1,64 +1,54 @@
 {
-    "name": "rd-frontend",
-    "version": "0.1.0",
-    "private": true,
-    "dependencies": {
-        "@apollo/client": "^3.1.3",
-        "@material-ui/core": "^4.11.0",
-        "@material-ui/icons": "^4.9.1",
-        "cross-fetch": "^3.0.5",
-        "graphql": "^14.7.0",
-        "graphql-tag.macro": "^2.0.1",
-        "loglevel": "^1.6.8",
-        "moment": "^2.27.0",
-        "react": "^16.13.1",
-        "react-datepicker": "^3.1.3",
-        "react-dom": "^16.13.1",
-        "react-dropzone": "^11.0.2",
-        "react-helmet": "^6.1.0",
-        "react-html-parser": "^2.0.2",
-        "react-infinite-scroller": "^1.2.4",
-        "react-router-dom": "^5.2.0",
-        "react-scripts": "3.4.1",
-        "styled-components": "^5.1.1",
-        "subscriptions-transport-ws": "^0.9.16"
-    },
-    "scripts": {
-        "start": "react-scripts start",
-        "build": "react-scripts build",
-        "test": "react-scripts test",
-        "eject": "react-scripts eject",
-        "format": "prettier --write .",
-        "intro": "node src/utils/introspection.js",
-        "lint": "eslint --fix ."
-    },
-    "eslintConfig": {
-        "extends": "react-app"
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
-    },
-    "devDependencies": {
-        "@types/react": "^16.9.45",
-        "@types/react-helmet": "^6.1.0",
-        "@types/react-html-parser": "^2.0.1",
-        "@types/uuid": "^8.0.1",
-        "eslint-config-airbnb": "18.2.0",
-        "eslint-config-prettier": "^6.11.0",
-        "eslint-plugin-import": "^2.21.2",
-        "eslint-plugin-jsx-a11y": "^6.3.0",
-        "eslint-plugin-react": "^7.20.5",
-        "eslint-plugin-react-hooks": "^2.0.0",
-        "prettier": "2.0.5",
-        "typescript": "^3.9.5"
-    }
+  "name": "rd-frontend",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@apollo/client": "^3.1.3",
+    "@material-ui/core": "^4.11.0",
+    "@material-ui/icons": "^4.9.1",
+    "cross-fetch": "^3.0.5",
+    "graphql": "^14.7.0",
+    "graphql-tag.macro": "^2.0.1",
+    "loglevel": "^1.6.8",
+    "moment": "^2.27.0",
+    "react": "^16.13.1",
+    "react-datepicker": "^3.1.3",
+    "react-dom": "^16.13.1",
+    "react-dropzone": "^11.0.2",
+    "react-helmet": "^6.1.0",
+    "react-html-parser": "^2.0.2",
+    "react-infinite-scroller": "^1.2.4",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "3.4.1",
+    "styled-components": "^5.1.1",
+    "subscriptions-transport-ws": "^0.9.16"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "intro": "node src/utils/introspection.js",
+    "lint": "prettier-standard --lint"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "@types/react": "^16.9.45",
+    "@types/react-helmet": "^6.1.0",
+    "@types/react-html-parser": "^2.0.1",
+    "@types/uuid": "^8.0.1",
+    "prettier-standard": "^16.4.1",
+    "typescript": "^3.9.5"
+  }
 }
