--- src/components/PostFeed.styles.js
@@ -1,16 +1,0 @@
-import styled from "styled-components";
-
-import LoginBackground from "../images/backgroundLogin.svg";
-
-const Banner = styled.div`
-    // Image at the top
-    padding: 100px 100px;
-    background: url(${LoginBackground});
-    background-color: #fffdfd;
-    opacity: 1;
-    background-size: 100%;
-    background-repeat: no-repeat;
-    border-radius: 20px;
-`;
-
-export { Banner };
