+++ src/utils/MajorMinor.json
@@ -1,5 +1,5 @@
 {
-    "majors": "Architecture;Architectural Studies;Bioengineering;Chemical and Biomolecular Engineering;Civil and Environmental Engineering;Computational and Applied Mathematics;Computer Science;Electrical and Computer Engineering;Mechanical Engineering;Materials Science and Nanoengineering;Statistics;Ancient Mediterranean Civilizations;Art History;Asian Studies;Classical Studies;English;European Studies;French Studies;German Studies;History;Latin American Studies; Medieval and Early Modern Studies;Philosophy;Religion;Spanish and Portuguese;Study of Women, Gender, and Sexuality;Visual and Dramatic Arts;Music;Astronomy;Astrophysics;Biochemistry and Cell Biology;Biological Sciences;Chemical Physics;Chemistry;Earth Science;Ecology and Evolutionary Biology;Environmental Science;Kinesiology;Mathematics;Neuroscience;Physics;Anthropology;Cognitive Sciences;Economics;Linguistics;Managerial Studies*;Mathematical Economic Analysis;Political Science;Psychology;Social Policy Analysis;Sociology;Sport Management",
-    "minors": "African Studies;Anthropology;Biochemistry and Cell Biology;Business;Cinema and Media Studies;Computational and Applied Mathematics;Data Science;Ecology and Evolutionary Biology;Energy and Water Sustainability;Engineering Design;Environmental Studies;Financial Computation and Modeling;Global Health Technologies;Jewish Studies;Mathematics;Medical Humanities;Museums and Cultural Heritage;Naval Science;Neuroscience;Physics;Politics, Law, and Social Thought;Poverty, Justice, and Human Capabilities;Religion;Sociology;Statistics;",
-    "colleges": "Will_Rice;Duncan;Baker;Martel;McMurtry;Brown;Jones;Hanszen;Wiess;Sid_Richardson;Lovett"
+  "majors": "Architecture;Architectural Studies;Bioengineering;Chemical and Biomolecular Engineering;Civil and Environmental Engineering;Computational and Applied Mathematics;Computer Science;Electrical and Computer Engineering;Mechanical Engineering;Materials Science and Nanoengineering;Statistics;Ancient Mediterranean Civilizations;Art History;Asian Studies;Classical Studies;English;European Studies;French Studies;German Studies;History;Latin American Studies; Medieval and Early Modern Studies;Philosophy;Religion;Spanish and Portuguese;Study of Women, Gender, and Sexuality;Visual and Dramatic Arts;Music;Astronomy;Astrophysics;Biochemistry and Cell Biology;Biological Sciences;Chemical Physics;Chemistry;Earth Science;Ecology and Evolutionary Biology;Environmental Science;Kinesiology;Mathematics;Neuroscience;Physics;Anthropology;Cognitive Sciences;Economics;Linguistics;Managerial Studies*;Mathematical Economic Analysis;Political Science;Psychology;Social Policy Analysis;Sociology;Sport Management",
+  "minors": "African Studies;Anthropology;Biochemistry and Cell Biology;Business;Cinema and Media Studies;Computational and Applied Mathematics;Data Science;Ecology and Evolutionary Biology;Energy and Water Sustainability;Engineering Design;Environmental Studies;Financial Computation and Modeling;Global Health Technologies;Jewish Studies;Mathematics;Medical Humanities;Museums and Cultural Heritage;Naval Science;Neuroscience;Physics;Politics, Law, and Social Thought;Poverty, Justice, and Human Capabilities;Religion;Sociology;Statistics;",
+  "colleges": "Will_Rice;Duncan;Baker;Martel;McMurtry;Brown;Jones;Hanszen;Wiess;Sid_Richardson;Lovett"
 }
